flag updating and note appending, #5232
