+++ public/src/client/topic/flag.js
@@ -48,7 +48,7 @@ define('forum/topic/flag', [], function () {
 		if (!pid || !reason) {
 			return;
 		}
-		socket.emit('posts.flag', {pid: pid, reason: reason}, function (err) {
+		socket.emit('flags.create', {pid: pid, reason: reason}, function (err) {
 			if (err) {
 				return app.alertError(err.message);
 			}
